refactor(#3109): consolidate 39 more local compileAndRun copies into tests/helpers/compile.ts (slice 2)
14 identical-body clusters (body-hash grouped) → 13 shared helpers; call
sites unchanged via import aliasing. Parity-proven: vitest --reporter=json
before/after over all 39 files — identical per-test result set (287 tests,
258 pass / 29 pre-existing fails). The one non-equivalence the hash missed
(issue-1594b/723-tdz's LOCAL reflected-stub buildImports, not src/runtime's)
was caught by the parity gate and moved verbatim into the helper
(reflectedStubImports). tsc + prettier clean. Zero src/ changes.

Running count: 58 of 132 local definitions removed (19 + 39); ~67 remain
(singletons — next slice needs semantic clustering).


✓
