+++ tests/issue-1024.test.ts
@@ -1,16 +1,5 @@
 import { describe, test, expect } from "vitest";
-import { compile } from "../src/index.ts";
-import { buildImports } from "../src/runtime.ts";
-
-async function compileAndRun(src: string): Promise<number> {
-  const r = await compile(src, { fileName: "test.ts" });
-  if (!r.success) {
-    throw new Error(`Compile error: ${r.errors[0]?.message}`);
-  }
-  const imports = buildImports(r.imports, undefined, r.stringPool);
-  const { instance } = await WebAssembly.instantiate(r.binary, imports);
-  return (instance.exports as Record<string, Function>).test() as number;
-}
+import { compileAndRunTestNumber as compileAndRun } from "./helpers/compile.js";
 
 describe("#1024 — Destructuring defaults with holes/undefined", () => {
   test("[x = 23] = [,] — hole triggers default", async () => {
