+++ tests/issue-1133.test.ts
@@ -1,14 +1,5 @@
 import { test, expect, describe } from "vitest";
-import { compile } from "../src/index.ts";
-import { buildImports } from "../src/runtime.ts";
-
-async function compileAndRun(src: string): Promise<any> {
-  const r = await compile(src, { fileName: "test.ts" });
-  if (!r.success) throw new Error(`Compile error: ${r.errors?.[0]?.message}`);
-  const imports = buildImports(r.imports, undefined, r.stringPool);
-  const instance = new WebAssembly.Instance(new WebAssembly.Module(r.binary), imports);
-  return (instance.exports as any).test();
-}
+import { compileAndRunTestSync as compileAndRun } from "./helpers/compile.js";
 
 describe("#1133 — any-typed string equality uses content comparison, not identity", () => {
   test("'hello' === 'hello' returns true for any-typed values", async () => {
