+++ tests/issue-1494.test.ts
@@ -8,26 +8,7 @@
 // `__dirname`, `__filename`, and `importMetaUrl`.
 
 import { describe, expect, it } from "vitest";
-import { compile } from "../src/index.js";
-import { buildImports as buildRuntimeImports } from "../src/runtime.js";
-
-async function compileAndRun(source: string, deps?: Record<string, unknown>): Promise<Record<string, Function>> {
-  const result = await compile(source);
-  if (!result.success) {
-    throw new Error(
-      `Compile failed:\n${result.errors.map((e) => `  L${e.line}: ${e.message}`).join("\n")}\nWAT:\n${result.wat}`,
-    );
-  }
-  if (!WebAssembly.validate(result.binary)) {
-    throw new Error(`Invalid Wasm binary\nWAT:\n${result.wat}`);
-  }
-  const runtimeResult = buildRuntimeImports(result.imports ?? [], deps, result.stringPool);
-  const { instance } = await WebAssembly.instantiate(result.binary, runtimeResult);
-  if (runtimeResult.setExports) {
-    runtimeResult.setExports(instance.exports as Record<string, Function>);
-  }
-  return instance.exports as Record<string, Function>;
-}
+import { compileAndRunRuntimeDeps as compileAndRun } from "./helpers/compile.js";
 
 describe("#1494 — __dirname / __filename / import.meta.url", () => {
   it("__dirname resolves to the loader-injected value", async () => {
