+++ tests/issue-2031.test.ts
@@ -1,22 +1,12 @@
 import { describe, it, expect } from "vitest";
 import { compile } from "../src/index.js";
-import { buildImports } from "../src/runtime.js";
+import { compileAndRunHost as compileAndRun } from "./helpers/compile.js";
 
 // #2031/#2032 — destructuring spec-conformance pair:
 //   - array destructuring with default + rest + a source shorter than the
 //     fixed bindings must NOT trap (array.copy source offset clamped)
 //   - computed-key object destructuring `{ [k]: v }` must bind the real value
 //     when the key is a compile-time-constant string (struct fast path)
-async function compileAndRun(source: string): Promise<Record<string, Function>> {
-  const result = await compile(source);
-  if (!result.success) {
-    throw new Error(`Compile failed:\n${result.errors.map((e) => `  L${e.line}: ${e.message}`).join("\n")}`);
-  }
-  const imports = buildImports(result.imports, undefined, result.stringPool);
-  const { instance } = await WebAssembly.instantiate(result.binary, imports as unknown as WebAssembly.Imports);
-  return instance.exports as Record<string, Function>;
-}
-
 describe("#2031 — array destructuring default + rest + short source", () => {
   it("const [p, q = 9, ...rest] = [1] binds p=1, q=9, rest=[]", async () => {
     const e = await compileAndRun(`export function test(): number {
