+++ tests/issue-2756.test.ts
@@ -1,7 +1,6 @@
 // Copyright (c) 2026 Loopdive GmbH. Licensed under Apache-2.0 WITH LLVM-exception.
 import { describe, expect, it } from "vitest";
-import { compile } from "../src/index.js";
-import { buildImports } from "../src/runtime.js";
+import { compileAndRunHost as compileAndRun } from "./helpers/compile.js";
 
 /**
  * #2756 — array-pattern element with an object-literal / class-expression default
@@ -25,16 +24,6 @@ import { buildImports } from "../src/runtime.js";
  * (§15.7.14 ClassDefinitionEvaluation defines static members AFTER
  * SetFunctionName, overriding `.name`).
  */
-
-async function compileAndRun(source: string): Promise<Record<string, Function>> {
-  const result = await compile(source);
-  if (!result.success) {
-    throw new Error(`Compile failed:\n${result.errors.map((e) => `  L${e.line}: ${e.message}`).join("\n")}`);
-  }
-  const imports = buildImports(result.imports, undefined, result.stringPool);
-  const { instance } = await WebAssembly.instantiate(result.binary, imports as unknown as WebAssembly.Imports);
-  return instance.exports as Record<string, Function>;
-}
 
 describe("#2756 — array-pattern object/class default null-deref", () => {
   it("object-literal default fires when element absent (was null-deref)", async () => {
