+++ tests/issue-280.test.ts
@@ -5,29 +5,7 @@
  * and various function expression patterns.
  */
 import { describe, it, expect } from "vitest";
-import { compile } from "../src/index.js";
-
-async function compileAndRun(source: string) {
-  const result = await compile(source);
-  expect(
-    result.success,
-    `Compile failed:\n${result.errors.map((e) => `  L${e.line}: ${e.message}`).join("\n")}\nWAT:\n${result.wat}`,
-  ).toBe(true);
-  const imports = {
-    env: {
-      console_log_number: () => {},
-      console_log_string: () => {},
-      console_log_bool: () => {},
-      __make_callback: () => {},
-    },
-  };
-  try {
-    const { instance } = await WebAssembly.instantiate(result.binary, imports);
-    return instance.exports as Record<string, Function>;
-  } catch (e) {
-    throw new Error(`Instantiation failed: ${e}\nWAT:\n${result.wat}`);
-  }
-}
+import { compileAndRunStubsCallback as compileAndRun } from "./helpers/compile.js";
 
 describe("Issue #280: Function expression name binding", () => {
   it("named function expression self-reference (recursion)", async () => {
