+++ tests/issue-862.test.ts
@@ -9,22 +9,7 @@
  * re-throw from .next() after yielded values are consumed.
  */
 import { describe, it, expect } from "vitest";
-import { compile } from "../src/index.ts";
-import { buildImports } from "../src/runtime.ts";
-
-async function compileAndRun(source: string): Promise<{ success: boolean; result?: number; error?: string }> {
-  const compiled = await compile(source, { fileName: "test.ts" });
-  if (!compiled.success) return { success: false, error: compiled.errors[0]?.message };
-  try {
-    const imports = buildImports(compiled.imports, undefined, compiled.stringPool);
-    const mod = new WebAssembly.Module(compiled.binary);
-    const inst = new WebAssembly.Instance(mod, imports);
-    const ret = (inst.exports as any).test();
-    return { success: true, result: ret };
-  } catch (e: any) {
-    return { success: false, error: `${e.constructor.name}: ${e.message}` };
-  }
-}
+import { compileAndRunResultObject as compileAndRun } from "./helpers/compile.js";
 
 describe("Issue #862: generator exception deferral", () => {
   it("generator throw is deferred to .next() and catchable", async () => {
