+++ tests/issue-864.test.ts
@@ -10,22 +10,7 @@
  * is coerced to externref, __box_symbol is used instead of __box_number.
  */
 import { describe, it, expect } from "vitest";
-import { compile } from "../src/index.ts";
-import { buildImports } from "../src/runtime.ts";
-
-async function compileAndRun(source: string): Promise<{ success: boolean; result?: number; error?: string }> {
-  const compiled = await compile(source, { fileName: "test.ts" });
-  if (!compiled.success) return { success: false, error: compiled.errors[0]?.message };
-  try {
-    const imports = buildImports(compiled.imports, undefined, compiled.stringPool);
-    const mod = new WebAssembly.Module(compiled.binary);
-    const inst = new WebAssembly.Instance(mod, imports);
-    const ret = (inst.exports as any).test();
-    return { success: true, result: ret };
-  } catch (e: any) {
-    return { success: false, error: `${e.constructor.name}: ${e.message}` };
-  }
-}
+import { compileAndRunResultObject as compileAndRun } from "./helpers/compile.js";
 
 describe("Issue #864: WeakMap/WeakSet Symbol keys", () => {
   it("Symbol as WeakMap key", async () => {
